requirements.txt
appdirs==1.4.0
argon2==0.1.10
argon2-cffi==16.3.0
cffi==1.9.1
click==6.7
dominate==2.3.1
enum34==1.1.6
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
MySQL-python==1.2.5
packaging==16.8
pycparser==2.17
pyparsing==2.1.10
PyYAML==3.12
six==1.10.0
SQLAlchemy==1.1.5
uWSGI==2.0.14
vex==0.0.18
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.11.15
